Use `URI.encode_www_form_component` instead of `URI.encode`
`URI.encode_www_form_component` is an implementation of
http://www.w3.org/TR/html5/forms.html#url-encoded-form-data
which is exactly what is needed to build the body of a request with
`Content-Type: application/x-www-form-urlencoded`.

`URI.encode` caused problems as for example e-mails containing a `+`sign
were not encoded correctly.

`email=test+plus@domain.com` is decoded to `email=test plus@domain.com`
by the recipient of the `POST`-request with
`Content-Type: application/x-www-form-urlencoded`.